requirements copy.txt
anyio==3.6.2
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
cmake==3.26.3
contourpy==1.0.7
cycler==0.11.0
dlib==19.24.1
dnspython==2.3.0
docopt==0.6.2
face-recognition==1.3.0
face-recognition-models==0.3.0
fastapi==0.95.0
fonttools==4.39.3
google-api-core==2.11.0
google-auth==2.17.0
google-cloud-core==2.3.2
google-cloud-pubsub==2.16.0
google-cloud-storage==2.5.0
google-cloud-vision==3.1.4
google-crc32c==1.5.0
google-resumable-media==2.4.1
googleapis-common-protos==1.59.0
grpc-google-iam-v1==0.12.6
grpcio==1.53.0
grpcio-status==1.53.0
h11==0.14.0
httptools==0.5.0
idna==3.4
importlib-resources==5.12.0
isodate==0.6.1
kiwisolver==1.4.4
matplotlib==3.7.1
numpy==1.24.2
opencv-python-headless==4.7.0.72
packaging==23.1
Pillow==9.3.0
pipreqs==0.4.11
proto-plus==1.22.2
protobuf==4.22.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.10.7
pymongo==4.3.3
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0
requests==2.28.2
rsa==4.9
six==1.16.0
sniffio==1.3.0
starlette==0.26.1
typing_extensions==4.5.0
urllib3==1.26.15
uvicorn==0.21.1
uvloop==0.17.0
watchfiles==0.19.0
websockets==10.4
yarg==0.1.9
zipp==3.15.0
